requirements.txt
discord.py>=2.0.0
motor>=3.0.0
pymongo>=4.0.0
python-dotenv>=0.19.0
python-dateutil>=2.8.2
asyncio>=3.4.3
colorama>=0.4.4
orjson>=3.6.0
dnspython>=2.1.0
aiohttp>=3.8.0
cryptography>=36.0.0
PyNaCl>=1.5.0
typing-extensions>=4.0.0
pytz>=2021.3
